Complete lib extraction
